+++ stat/stats-server/src/main/resources/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS STATS
 (
-    ID   int4 generated always as identity primary key,
+    ID   BIGINT generated always as identity primary key,
     APP  varchar(256) NOT NULL,
     URI  varchar,
     IP   varchar,
